Adding code to scan device folder on launch and populate the GroupFile list in the main activity

+++ app/src/main/java/com/csus/csc258/csc258_group_project/GroupFile.java
@@ -1,37 +1,39 @@
 package com.csus.csc258.csc258_group_project;
 
 import android.content.Context;
-import android.provider.Settings;
 import android.util.Log;
 import android.view.View;
 import android.widget.Toast;
-
 import java.io.File;
 import java.io.FileWriter;
 
 /**
- * Created by bharr on 3/1/2016.
+ * Representation of a file that is owned by a group
  */
     public class GroupFile
     {
         private String mFileName;
         private String mPath;
-        private MainActivity mActivity;
+        private Context mContext;
 
         // The id of the file
         private int mFId;
 
+        private static final String TAG = "GroupFile";
+
         /**
          * Create a new file object
          * @param name The name of the file
-         * @param activity The main activity window (used for displaying Toast messages)
+         * @param context The application context (used for toast messages)
+         * @param deviceID The name of the device (for subfolder creation)
          */
-        public GroupFile(String name, MainActivity activity) {
+        public GroupFile(String name, Context context, String deviceID) {
             mFileName = name;
-            mActivity = activity;
-            mPath = mActivity.getFilesDir().getAbsolutePath() + File.separator + getDeviceId(mActivity);
+            mContext = context;
+            mPath = context.getFilesDir().getAbsolutePath() + File.separator + deviceID;
             mFId = View.generateViewId();
-            writeFileOnInternalStorage("Sample Body");
+            if (!fileExistsInternalStorage())
+                writeFileOnInternalStorage("Sample Body");
         }
 
         /**
@@ -51,71 +53,81 @@ import java.io.FileWriter;
         public int getFileId() { return mFId; }
 
         /**
-         * Deletes the file on the file system
+         * Returns true if the file exists on internal storage
+         * @return True if the file exists on internal storage, false otherwise
          */
-        public void deleteFile() {
-            deleteFileFromInternalStorage();
+        public boolean fileExistsInternalStorage() {
+            File f = new File(mPath, mFileName);
+            return f.exists();
         }
 
-        private void writeFileOnInternalStorage(String sBody){
-            //File samplefiledir = new File(mcoContext.getFilesDir(),"sampledata");
-            File samplefiledir = new File(mPath);
+        /**
+         * Deletes the file on the file system
+         * @return True if successful, false otherwise
+         */
+        public boolean deleteFile() {
+            if (fileExistsInternalStorage())
+                return deleteFileFromInternalStorage();
+            else
+                return false;
+        }
+
+        /**
+         * Writes the file to the internal storage
+         * @param sBody The text content of the file
+         * @return True if successful, false otherwise
+         */
+        private boolean writeFileOnInternalStorage(String sBody){
+            File fileDir = new File(mPath);
             Boolean write_successful = false;
-            if(!samplefiledir.exists()){
-                samplefiledir.mkdir();
+            if(!fileDir.exists()){
+                fileDir.mkdir();
             }
             try{
-                File samplefile = new File(samplefiledir, mFileName);
+                File samplefile = new File(fileDir, mFileName);
                 FileWriter writer = new FileWriter(samplefile);
                 writer.append(sBody);
                 writer.flush();
                 writer.close();
                 write_successful = true;
-                Toast.makeText(mActivity.getApplicationContext(), "Sample File Created Successfully in : " + samplefiledir + "\t with the name:\t" + mFileName, Toast.LENGTH_SHORT).show();
+                Toast.makeText(mContext, "Sample File Created Successfully in : " + fileDir + "\t with the name:\t" + mFileName, Toast.LENGTH_SHORT).show();
             }
             catch (Exception e){
-                Toast.makeText(mActivity.getApplicationContext(), e.getMessage(), Toast.LENGTH_SHORT).show();
-                Log.e("ERROR:---", "Could not write file" + e.getMessage());
-                write_successful = false;
+                Toast.makeText(mContext, e.getMessage(), Toast.LENGTH_SHORT).show();
+                Log.e(TAG, "Could not write file" + e.getMessage());
             }
+
+            return write_successful;
         }
 
-        private void deleteFileFromInternalStorage(){
-            //File samplefiledir = new File(mcoContext.getFilesDir(),"sampledata");
-            File samplefiledir = new File(mPath);
+        private boolean deleteFileFromInternalStorage(){
+            File fileDir = new File(mPath);
             Boolean isDeleted = null;
             File f = new File(mPath,mFileName);
             try {
                 if(f.exists()) {
                     f.delete();
                     isDeleted = true;
-                    Toast.makeText(mActivity.getApplicationContext(), "Sample File Deleted Successfully from : " + samplefiledir + "\t with the name:\t" + mFileName, Toast.LENGTH_SHORT).show();
+                    Toast.makeText(mContext, "Sample File Deleted Successfully from : " + fileDir + "\t with the name:\t" + mFileName, Toast.LENGTH_SHORT).show();
                 }
                 else
                 {
                     isDeleted = false;
-                    Toast.makeText(mActivity.getApplicationContext(), "Could not delete file", Toast.LENGTH_SHORT).show();
+                    Toast.makeText(mContext, "Could not delete file", Toast.LENGTH_SHORT).show();
                 }
             }
             catch (Exception e){
-                Toast.makeText(mActivity.getApplicationContext(), e.getMessage(), Toast.LENGTH_SHORT).show();
+                Toast.makeText(mContext, e.getMessage(), Toast.LENGTH_SHORT).show();
                 isDeleted = false;
-                Log.e("ERROR:---", "Could not delete file" + e.getMessage());
+                Log.e(TAG, "Could not delete file" + e.getMessage());
             }
+
+            return isDeleted;
         }
 
         //TODO
         private void renameGroupFile(String sFilename){
-            Context mcoContext = mActivity.getApplicationContext();
-            //File samplefiledir = new File(mcoContext.getFilesDir(),sFilename);
-            String device_id = getDeviceId(mcoContext);
-            File samplefiledir = new File(mcoContext.getFilesDir().getAbsolutePath() + File.separator + device_id);
-            samplefiledir.renameTo(samplefiledir);
-        }
-
-        private String getDeviceId (Context context)
-        {
-            String device_id = Settings.Secure.getString(context.getContentResolver(), Settings.Secure.ANDROID_ID);
-            return device_id;
+            File fileDir = new File(mPath);
+            fileDir.renameTo(fileDir);
         }
     }
